+++ src/server/message_handler/completion/blank_node_property.rs
@@ -20,7 +20,7 @@ pub(super) async fn completions(
         server_rc,
         &environment,
         template_context,
-        CompletionTemplate::PredicateCompletion,
+        CompletionTemplate::PredicateCompletionContextInsensitive,
         true,
     )
     .await