+++ src/server/message_handler/completion/predicate.rs
@@ -30,7 +30,7 @@ pub(super) async fn completions(
             server_rc_1,
             &environment_1,
             template_context_1,
-            CompletionTemplate::PredicateCompletion,
+            CompletionTemplate::PredicateCompletionContextInsensitive,
             true,
         )
         .await